Remove nodiscard V8 configuration option
Nodiscard option means that the return value of a function is assigned to
a variable and not omitted. Typically, this is used to enforce return
status checking in the code of the client. Otherwise programmers can
just ignore the return value for operations that they assume to always
succeed.

Unfortunately this option is breaking old code. Old code needs to be
fixed before this can be re-enabled.

Signed-off-by: Markku Kylänpää <markku.kylanpaa@vtt.fi>